Fix tms9918a transparent color rendering
Currently transparent color (index 0) is always rendered black -- this
is wrong, transparent color should be a copy of the TMS9918A
background color control register value or if that's zero, black. This
patch fixes the problem by reassigning the RGB value for the
transparent color always before the screen is rendered.

Signed-off-by: Juha Riihimäki <juha.riihimaki@nokia.com>